backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.0.1",
    "fastapi>=0.128.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
]

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true
